--- apparent/config.py
@@ -1,2 +1,0 @@
-URL = "https://apparent.topology.rocks/us_physician_referral_networks.csv"
-query_file = "query.sql"
