+++ apparent/networks/compare.py
@@ -1,0 +1,11 @@
+# TODO: KILT or other comparison methods
+
+
+class NetworkComparison:
+    def __init__(self, networks):
+        self.networks = networks
+
+    def compare(self):
+        # Compare the two networks
+        # and return the result
+        pass
