+++ apparent/networks/describe.py
@@ -1,32 +1,22 @@
-# Build Graphs
-def build_network(edges_df, hsanum, year):
-    """Create a networkx graph given a dataframe of edges, a region, and year."""
+import networkx as nx
 
-    # initialize an undirected graph G
-    G = nx.Graph()
+# KILT
+import sys
 
-    # populate G
-    G = nx.from_pandas_edgelist(
-        df=edges_df[(edges_df.hsanum == hsanum) & (edges_df.year == year)],
-        source="npi_a",
-        target="npi_b",
-        edge_attr=["a2b", "b2a"],
-    )
-
-    # sanity check
-    assert G.is_directed() is False
-    assert G.is_multigraph() is False
-
-    return G
+sys.path.append("/Users/jeremy.wayland/repos/dev/curvature-filtrations")
+from curvature_filtrations.kilt import KILT, CURVATURE_MEASURES
 
 
-def process_row(row):
-    A, nodes, curvature = build_network(edges_df, row["hsanum"], row["year"])
-    return pd.Series({"adjacency": A, "nodes": nodes, "curvature": curvature})
+class NetworkDescriber:
+
+    # TODO: Curvature, NetworkX, Cycle Representatives with OatPython
+    pass
 
 
 # Compute Features
-def compute_features(G, node_features, edge_features):
+def compute_features(G, node_features, edge_features, **kwargs):
+
+    # TODO: Need to figure out how to pass curvature specific measures to KILT
     """Compute features for a given graph.
 
     This function computes features for a given graph. The features
@@ -62,6 +52,7 @@ def compute_features(G, node_features, edge_features):
             features[feature] = nx.closeness_centrality(G)
         elif feature == "pagerank":
             features[feature] = nx.pagerank(G)
+
         else:
             raise ValueError(f"Unknown node feature: {feature}")
 
@@ -69,6 +60,14 @@ def compute_features(G, node_features, edge_features):
     for feature in edge_features:
         if feature == "edge_betweenness":
             features[feature] = nx.edge_betweenness_centrality(G)
+
+        ## CURVATURE
+        elif feature in CURVATURE_MEASURES:
+            # TODO: CHECK kwargs and pass proper ones to KILT
+            kilt = KILT(measure=feature, **kwargs)
+            kilt.fit(G)
+            # Probably a nice getter from nx to get edge/feature dict
+            features[feature] = dict(zip(G.edges(), kilt.curvature))
         else:
             raise ValueError(f"Unknown edge feature: {feature}")
